Allow parsing with HTML5
Nokogiri is on the path to parsing with HTML5 by default:
https://github.com/sparklemotion/nokogiri/issues/2331

But, there are some things they still need to do. For those of us who
want to opt-in to HTML5 parsing, I've added an option for it. This will
prevent the gem from messing with the structure of the html
(specifically, prematurely closing <a> tags that wrapped table elements.